requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Pydantic 버전 호환성을 위해 특정 버전 사용
pydantic==2.8.2
pydantic-core==2.20.1

# LangChain 관련 패키지 (호환되는 버전)
langchain==0.2.16
langchain-google-genai==1.0.10
langchain-community==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4

# 벡터 데이터베이스 및 임베딩
chromadb==0.5.5
sentence-transformers==3.0.1

# 기타 필수 패키지
python-multipart==0.0.6
numpy==1.24.3
scikit-learn==1.3.2
python-dotenv==1.0.0

# 문서 처리
pypdf==3.17.4
python-docx==1.1.0

# 검색 및 순위
rank-bm25==0.2.2

# 진행 표시줄 및 UI
tqdm==4.66.1
ipywidgets==8.1.1